feat: implemented suggestions in pr
